+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.6.46] - 2025-03-11
+
+### Changed
+
+- Update api `Affise.getReferrerOnServer` moved to `Affise.getDeferredDeeplink`.
+- Update api `Affise.getReferrerOnServerValue` moved to `Affise.getDeferredDeeplinkValue`.
+
 ## [1.6.45] - 2025-02-25
 
 ### Added
@@ -151,6 +158,7 @@
 - Improve `AffiseModuleManager`
 - Update for demo app
   
+[1.6.46]: https://github.com/affise/sdk-ios/compare/1.6.45...1.6.46
 [1.6.45]: https://github.com/affise/sdk-ios/compare/1.6.44...1.6.45
 [1.6.44]: https://github.com/affise/sdk-ios/compare/1.6.43...1.6.44
 [1.6.43]: https://github.com/affise/sdk-ios/compare/1.6.42...1.6.43